Change namespace to be separated by UID2 and EUID

+++ scripts/eks/stop_eks_enclave.sh
@@ -1,2 +1,7 @@
-kubectl delete namespace compute --ignore-not-found=true
+if [ -z "${IDENTITY_SCOPE}" ]; then
+  echo "IDENTITY_SCOPE can not be empty"
+  exit 1
+fi
+
+kubectl delete namespace ${IDENTITY_SCOPE,,} --ignore-not-found=true
 kubectl delete namespace nitro-enclaves --ignore-not-found=true
